Move HostLink to common
