+++ assets/js/common/HostLink/index.js
@@ -1,0 +1,3 @@
+import HostLink from './HostLink';
+
+export default HostLink;
